+++ src/common_traits/deref_derefmut.rs
@@ -14,28 +14,3 @@ impl<T, P: PinnedVec<T>> DerefMut for ImpVec<T, P> {
         self.pinned_mut()
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn deref() {
-        let vec = ImpVec::new();
-        vec.imp_extend_from_slice(&['a', 'b', 'c']);
-
-        let pinned_deref = vec.deref();
-        assert_eq!(pinned_deref, vec.pinned_mut());
-    }
-
-    #[test]
-    fn deref_mut() {
-        let mut vec = ImpVec::new();
-        vec.imp_extend_from_slice(&['a', 'b', 'c']);
-
-        let pinned_deref = vec.deref_mut();
-        pinned_deref.push('d');
-
-        assert_eq!('d', vec[3]);
-    }
-}
