+++ src/common_traits/from.rs
@@ -9,18 +9,3 @@ impl<T, P: PinnedVec<T>> From<P> for ImpVec<T, P> {
         }
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use orx_split_vec::prelude::*;
-
-    #[test]
-    fn from() {
-        let mut splitvec = SplitVec::new();
-        splitvec.extend_from_slice(&['a', 'b', 'c']);
-
-        let impvec = ImpVec::from(splitvec);
-        assert_eq!(*impvec, &['a', 'b', 'c']);
-    }
-}
