+++ src/common_traits/tests/mod.rs
@@ -3,6 +3,3 @@ mod deref_derefmut;
 mod eq;
 mod from;
 mod into_iter;
-
-#[cfg(test)]
-mod tests;
